# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Projetlvre",
        "description": "une sorte de biblio en ligne",
        "version": "1.0"
    },
    "paths": {
...
        "/api/createlivre": {
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Create a new book",
                "description": "Allow a seller to upload and publish a new book.",
                "operationId": "e5a865203ff0694114b7bb171f0d8750",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nomL",
                                    "categorieL",
                                    "description",
                                    "path",
                                    "statutL",
                                    "date",
                                    "prixL"
                                ],
                                "properties": {
                                    "nomL": {
                                        "type": "string",
                                        "example": "The Art of Programming"
                                    },
                                    "categorieL": {
                                        "description": "Category ID",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A complete guide to modern programming concepts."
                                    },
                                    "path": {
                                        "description": "PDF file of the book",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "statutL": {
                                        "description": "1 = published, 0 = draft",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-03-22"
                                    },
                                    "prixL": {
                                        "type": "number",
                                        "example": 2500
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Book successfully created"
                                        },
                                        "book": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/livrecategorie/{categorie}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Retrieve books belonging to a specific category",
                "description": "Get books by category",
                "operationId": "f5b50e9002c595a8829acc587d6aa010",
                "parameters": [
                    {
                        "name": "categorie",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Books filtered by category retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "books": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Books filtered by category successfully retrieved"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}